+++ rich/jupyter.py
@@ -1,89 +1,82 @@
 import io
-from typing import Any, IO, Union
+from typing import Any, Iterable, IO, List, Optional, TYPE_CHECKING, Union
 
-from .console import Console as BaseConsole
+# from .console import Console as BaseConsole
+from .__init__ import get_console
+from .segment import Segment
 from .style import Style
+from .terminal_theme import DEFAULT_TERMINAL_THEME
+
+if TYPE_CHECKING:
+    from .console import Console, RenderableType
 
 JUPYTER_HTML_FORMAT = """\
-<pre style="white-space:pre;overflow-x:auto;line-height:1em;font-family:Menlo,'DejaVu Sans Mono',consolas,'Courier New',monospace">{code}</pre>
+<pre style="white-space:pre;overflow-x:auto;line-height:normal;font-family:Menlo,'DejaVu Sans Mono',consolas,'Courier New',monospace">{code}</pre>
 """
 
 
 class JupyterRenderable:
     """A shim to write html to Jupyter notebook."""
 
-    def __init__(self, text: str, html: str) -> None:
-        self.text = text
+    def __init__(self, html: str) -> None:
         self.html = html
 
-    def __str__(self) -> str:
-        return self.text
+    @classmethod
+    def render(self, rich_renderable: "RenderableType") -> str:
+        console = get_console()
+        segments = console.render(rich_renderable, console.options)
+        html = _render_segments(segments)
+        return html
 
     def _repr_html_(self) -> str:
         return self.html
 
 
-class Console(BaseConsole):
-    def __init__(self, **kwargs) -> None:
-        kwargs["file"] = io.StringIO()
-        kwargs["record"] = True
-        if "width" not in kwargs:
-            kwargs["width"] = 100
-        super().__init__(**kwargs)
+class JupyterMixin:
+    """Add to an Rich renderable to make it render in Jupyter notebook."""
 
-    def is_terminal(self) -> bool:
-        return True
-
-    def print(  # type: ignore
-        self,
-        *objects: Any,
-        sep=" ",
-        end="\n",
-        file: IO[str] = None,
-        style: Union[str, Style] = None,
-        emoji: bool = None,
-        markup: bool = None,
-        highlight: bool = None,
-        flush: bool = False
-    ) -> JupyterRenderable:
-        r"""Print to the console.
-
-        Args:
-            objects (positional args): Objects to log to the terminal.
-            sep (str, optional): String to write between print data. Defaults to " ".
-            end (str, optional): String to write at end of print data. Defaults to "\n".
-            style (Union[str, Style], optional): A style to apply to output. Defaults to None.
-            emoji (Optional[bool], optional): Enable emoji code, or ``None`` to use console default. Defaults to None.
-            markup (Optional[bool], optional): Enable markup, or ``None`` to use console default. Defaults to None
-            highlight (Optional[bool], optional): Enable automatic highlighting, or ``None`` to use console default. Defaults to None.
-        """
-        if file is None:
-            super().print(
-                *objects,
-                sep=sep,
-                end=end,
-                style=style,
-                emoji=emoji,
-                markup=markup,
-                highlight=highlight,
-            )
-        else:
-            Console(file=file).print(
-                *objects,
-                sep=sep,
-                end=end,
-                style=style,
-                emoji=emoji,
-                markup=markup,
-                highlight=highlight,
-            )
-
-        html = self.export_html(code_format=JUPYTER_HTML_FORMAT, inline_styles=True)
-        text = self.file.getvalue()  # type: ignore
-        self.file = io.StringIO()
-        jupyter_renderable = JupyterRenderable(text, html)
-        return jupyter_renderable
+    def _repr_html_(self) -> str:
+        console = get_console()
+        segments = list(console.render(self, console.options))  # type: ignore
+        html = _render_segments(segments)
+        return html
 
 
-console = Console()
-print = console.print
+def _render_segments(segments: Iterable[Segment]) -> str:
+    def escape(text: str) -> str:
+        """Escape html."""
+        return text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;")
+
+    fragments: List[str] = []
+    append_fragment = fragments.append
+    theme = DEFAULT_TERMINAL_THEME
+    for text, style, is_control in Segment.simplify(segments):
+        if is_control:
+            continue
+        text = escape(text)
+        if style:
+            rule = style.get_html_style(theme)
+            text = f'<span style="{rule}">{text}</span>' if rule else text
+            if style.link:
+                text = f'<a href="{style.link}">{text}</a>'
+        append_fragment(text)
+
+    code = "".join(fragments)
+    html = JUPYTER_HTML_FORMAT.format(code=code)
+
+    return html
+
+
+def display(segments: Iterable[Segment]) -> None:
+    """Render segments to Jupyter."""
+    from IPython.display import display as ipython_display
+
+    html = _render_segments(segments)
+    jupyter_renderable = JupyterRenderable(html)
+    ipython_display(jupyter_renderable)
+
+
+def print(*args, **kwargs) -> None:
+    """Proxy for Console print."""
+    console = get_console()
+    return console.print(*args, **kwargs)